Synchronously removing expired report data. WIP #20 (#116)
This change will synchronously remove expired report data when issuing a DEL /api/v1/expired_data when masu is in Debug mode. This is to demonstrate that removal of report data is successful given the retention policy. The default retention policy is 3 months of data and the policy can be configured with the environment variable MASU_RETAIN_NUM_MONTHS.